requirements.txt
streamlit
ollama
openai
google-generativeai
youtube-transcript-api
reportlab
python-dotenv
